Standalone modus mit Interval
 Fixes #24
