+++ .env
@@ -4,4 +4,4 @@ DATABASE_URL="file:./data.sqlite"
 PORT=80
 PRISMA_STUDIO_PORT=17523
 
-DISPLAY_MODE="default"
+OPERATION_MODE="standalone"
