--- views/display/default.pug
@@ -1,33 +1,0 @@
-extends /base.pug
-
-append variables
-    - baseTitle = "Display - Standard"
-
-append scripts 
-    script(src="/js/websocket.js")
-
-append styles
-    link(rel="stylesheet" href="/css/display.css")
-
-append content 
-    #last-entry.card
-        .last-entry-details
-            #totalTime #{lastRun.time.formattedDriveTime} #{(lastRun.penalty!=null) ? "(+" + lastRun.penalty + "s)" : ""}
-        .value #{lastRun.driver.firstName} - #{lastRun.driver.drivingClass} - #{lastRun.vehicle.name}
-        .value #{lastRun.notes}
-
-    #last-times-driver.card
-        table 
-            tr 
-                th Fahrzeit 
-                th Strafsekunden
-                th Gesamtzeit
-                th Fahrzeug
-                th Notiz
-            for run in allDriverRuns 
-                tr
-                    td #{run.time.formattedDriveTime}
-                    td #{run.penalty}
-                    td #{run.time.formattedTotalTime}
-                    td #{run.vehicle.name}
-                    td #{run.notes}
