--- views/display/empty.pug
@@ -1,13 +1,0 @@
-extends /base.pug
-
-append variables
-    - baseTitle = "Display - NO DATA"
-
-append scripts 
-    script(src="/js/websocket.js")
-
-append styles
-    link(rel="stylesheet" href="/css/display.css")
-
-append content 
-    h1 Keine Daten vorhanden!
