--- views/display/ranking.pug
@@ -1,33 +1,0 @@
-extends /base.pug
-
-append variables
-    - baseTitle = "Display - Ranking"
-
-append scripts 
-    script(src="/js/websocket.js")
-
-append styles
-    link(rel="stylesheet" href="/css/display.css")
-
-append content 
-    h1 Rangliste
-    table
-        tr
-            th Platz
-            th Fahrer
-            th Klasse
-            th Fahrzeit
-            th Strafe 
-            th Gesamtzeit
-            th Fahrzeug
-            th Notiz
-        for run in bestTimes
-            tr
-                td #{run.position}
-                td #{run.driver.firstName} #{run.driver.lastName}
-                td #{run.driver.drivingClass}
-                td #{run.time.formattedDriveTime}
-                td #{run.penalty}
-                td #{run.time.formattedTotalTime}
-                td #{run.vehicle.name}
-                td #{run.notes}
